Refactor Additional Task 3
